Rename: auth -> member로 도메인 변경
