+++ src/main/java/semohan/semohan/domain/member/dto/SmsVerificationDto.java
@@ -1,4 +1,4 @@
-package semohan.semohan.domain.auth.dto;
+package semohan.semohan.domain.member.dto;
 
 import jakarta.validation.constraints.Pattern;
 import lombok.AllArgsConstructor;
